updated task solution
