+++ src/Root.tsx
@@ -1,0 +1,26 @@
+import { HashRouter, Routes, Route, Navigate } from 'react-router-dom';
+import { App } from './App';
+import { TabsProvider } from './context/TabsProvider';
+import { HomePage } from './components/HomePage';
+import { TabsPage } from './components/TabsPage/TabsPage';
+import { NotFoundPage } from './components/NotFoundPage';
+
+export const Root = () => (
+  <TabsProvider>
+    <HashRouter>
+      <Routes>
+        <Route path="/home" element={<Navigate to="/" replace />} />
+        <Route path="/" element={<App />}>
+          <Route index element={<HomePage />} />
+
+          <Route path="tabs">
+            <Route index element={<TabsPage />} />
+            <Route path=":tabId" element={<TabsPage />} />
+          </Route>
+
+          <Route path="*" element={<NotFoundPage />} />
+        </Route>
+      </Routes>
+    </HashRouter>
+  </TabsProvider>
+);
