+++ src/index.tsx
@@ -1,9 +1,4 @@
 import { createRoot } from 'react-dom/client';
-import { HashRouter } from 'react-router-dom';
-import { App } from './App';
+import { Root } from './Root';
 
-createRoot(document.getElementById('root') as HTMLElement).render(
-  <HashRouter>
-    <App />
-  </HashRouter>,
-);
+createRoot(document.getElementById('root') as HTMLElement).render(<Root />);
